+++ hoj-springboot/JudgeServer/src/main/java/top/hcode/hoj/judge/task/TestJudge.java
@@ -80,7 +80,7 @@ public class TestJudge extends AbstractJudge {
         } else if (sandBoxRes.getExitCode() != 0) {
             result.set("status", Constants.Judge.STATUS_RUNTIME_ERROR.getStatus());
             if (sandBoxRes.getExitCode() < 32) {
-                errMsg.append(String.format("ExitCode: %s (%s)\n", sandBoxRes.getExitCode(), SandboxRun.signals.get(sandBoxRes.getExitCode())));
+                errMsg.append(String.format("ExitCode: %s (%s)\n", sandBoxRes.getExitCode(), SandboxRun.signals.get(sandBoxRes.getExitCode().intValue())));
             } else {
                 errMsg.append(String.format("ExitCode: %s\n", sandBoxRes.getExitCode()));
             }
